refactor: rename model package to domain
- 이미 package를 model과 view로 나눠놓았기에, 그대로 model package이름만 domain으로 rename하였습니다.
- Step4에서 이미 테스트 가능한 부분(=domin package)에 대한 테스트는 끝나있습니다.
- Random값 생성부분은, RandomValueGenerator를 따로 분리하여 테스트되어 있습니다.
