+++ src/main/java/racingcar/domain/CarFactory.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 public final class CarFactory {
