+++ src/main/java/racingcar/domain/MoveCondition.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 @FunctionalInterface
 public interface MoveCondition {
