+++ src/main/java/racingcar/domain/MovePolicy.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 @FunctionalInterface
 public interface MovePolicy {
