+++ src/main/java/racingcar/domain/OneStepMovePolicy.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 public class OneStepMovePolicy implements MovePolicy {
