+++ src/main/java/racingcar/domain/RandomMoveCondition.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 @FunctionalInterface
 public interface RandomMoveCondition extends MoveCondition {
