+++ src/main/java/racingcar/domain/RandomValueGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 @FunctionalInterface
 public interface RandomValueGenerator {
