+++ src/main/java/racingcar/domain/SixtyPercentRandomMoveCondition.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 public class SixtyPercentRandomMoveCondition implements RandomMoveCondition {
